refactor: share generator module
